# configs/swagger/v1.yaml
swagger: "2.0"
info:
  description: "悟空聊天IM API"
  version: "1.0.0"
  title: "悟空聊天IM API"
host: "49.235.59.182:8080"
...
schemes:
  - "http"
basePath: "/v1"
paths:
...
  /user/login:
    post:
      tags:
        - "user"
      summary: "用户登录"
      description: "用户登录"
      operationId: "login"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "req"
          description: "登录请求"
          required: true
          schema:
            $ref: "#/definitions/UserLoginReq"
      responses:
        200:
          description: "返回"
          schema:
            $ref: '#/definitions/UserLoginResp'
        400:
          description: '错误'
          schema:
            $ref: '#/definitions/response'
...
  /user/devices:
    get:
      tags:
        - "user"
      summary: "获取用户登录设备"
      description: "获取用户登录设备"
      operationId: "devices"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        200:
          description: "返回"
          schema:
            type: array
            items:
              properties:
               device_id:
                type: string
                description: "设备ID"
               device_name:
                type: string
                description: "设备名称"
               device_model:
                type: string
                description: "设备类型"
               last_login:
                type: string
                description: "最后一次登录时间"
               self:
                type: integer
                description: "是否是本机 1:是0：否"
        400:
          description: '错误'
          schema:
            $ref: '#/definitions/response'
      security:
        - token: []
...
  /user/blacklists:
    get:
      tags:
        - 'user'
      summary: "用户黑名单列表"
      description: "用户黑名单列表"
      operationId: "blacklists"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        200:
          description: "成功"
          schema:
            type: array
            items:
              properties:
               uid:
                type: string
                description: "用户ID"
               name:
                type: string
                description: "用户名称"
               username:
                type: string
                description: '用户名称'
        400:
          description: '错误'
          schema:
            $ref: '#/definitions/response'
      security:
        - token: []
...
securityDefinitions:
  token:
    type: "apiKey"
    in: "header"
    name: "token"
    description: "用户token"
definitions:
...
  UserLoginReq:
    type: "object"
    properties:
      username:
        type: "string"
        description: "用户名"
      password:
        type: "string"
        description: "密码"
      name:
        type: "string"
        description: "名字"
      flag:
        type: integer
        format: int
        description: "设备标示 0.APP 1.PC"  
      device:
        type: object
        properties:
         device_id:
          type: string
          description: "设备ID"
         device_name:
          type: string
          description: "设备名称"
         device_model:
          type: string
          description: "设备型号"
  UserLoginResp:
    type: "object"
    properties:
      uid:
        type: "string"
        description: "用户唯一ID"
      token:
        type: "string"
        description: "用户凭证"
      username:
        type: "string"
        description: "登录用户名"
      name:
        type: "string"
        description: "用户名称"
      sex:
        type: integer
        description: "性别"
      category:
        type: string
        description: "用户分类 '客服'"
      short_no:
        type: string
        description: "用户唯一短编号"
      zone:
        type: string
        description: "手机区号"
      phone:
        type: string
        description: "手机号"
      setting:
        type: object
        description: "用户设置"
        properties:
          search_by_phone:
            type: integer
            description: "是否可以通过手机号搜索0.否1.是"
          search_by_short:
            type: integer
            description: "是否可以通过短编号搜索0.否1.是"
          new_msg_notice:
            type: integer
            description: "新消息通知0.否1.是"
          msg_show_detail:
            type: integer
            description: "显示消息通知详情0.否1.是"
          voice_on:
            type: integer
            description: "声音0.否1.是"
          shock_on:
            type: integer
            description: "震动0.否1.是"
  response:
    type: "object"
    properties:
      status:
        type: integer
        format: int
      msg:
        type: "string"